+++ engine/src/test/java/examples/packageWithFailings/ContainerWithFailingConstructor.java
@@ -1,0 +1,13 @@
+package examples.packageWithFailings;
+
+import net.jqwik.api.*;
+
+public class ContainerWithFailingConstructor {
+
+	public ContainerWithFailingConstructor() {
+		throw new RuntimeException("failing constructor");
+	}
+
+	@Example
+	void success() {}
+}